Enable dev server mode for the vault
As a first pass let's defer a more thorough analysis of the tradeoffs
around setting up the initial vault to a later point in time.
We enable the dev mode for vault for now so it can be usable for the
other bits in the pattern. We'll need to document the limits of
this and/or find a different tradeoff later on.